Fixing tests!
