+++ tests/test_path_searching.py
@@ -1,105 +1,7 @@
 import pytest
 
-from utils import read_warehouse_map
-from utils import neighbors
-from utils import a_star as shortest_path
-
-from logic import Direction
-from logic import Robot
-from logic import Magazine
-
-
-@pytest.fixture
-def warehouse_map1(tmpdir):
-    content = """22223
-1a1b3
-1c1d3
-1e1f3
-14449
-"""
-    file_ = tmpdir.join("warehouse.map")
-    file_.write(content)
-    return read_warehouse_map(str(file_), use_numpy=False)
-
-
-@pytest.mark.frontend
-def test_direction_class():
-    direction = Direction()
-    assert direction.get_up() is False
-    direction.set_up(True)
-    assert direction.get_up() is True
-    direction.set_down(True)
-    assert direction.get_down() is True
-    assert direction.get_direction_list() == [True, False, True, False]
-
-
-@pytest.mark.frontend
-def test_robot_class():
-    robot = Robot()
-    direction_list = robot.get_direction().get_direction_list()
-    assert direction_list == [False, False, False, False]
-    robot.set_y(4)
-    assert robot.get_y() == 4
-    robot.set_direction(Direction(up=True, left=True))
-    assert robot.get_direction().get_direction_list() == [True, False, False,
-                                                          True]
-
-
-@pytest.mark.frontend
-def test_magazine_class():
-    magazine = Magazine(5, 6)
-    magazine.show()
-    robot_array = ([[0, 1, 2, None],
-                   [3, 2, 1, None]])
-    arr = (
-        [
-            [2, 2, 2, 2, 3],
-            [1, "a", 1, "B", 3],
-            [1, "c", 1, "d", 3],
-            [1, "e", 1, "f", 3],
-            [1, "g", "h", "i", 3],
-            [1, 4, 4, 4, 9],
-        ]
-    )
-    magazine.update(arr, robot_array)
-    arr = (
-        [
-            [2, 2, 2, 2, 3],
-            [1, "a", 1, "b", 3],
-            [1, "c", 1, "d", 3],
-            [1, "e", 1, "f", 3],
-            [1, "g", 4, "i", 3],
-            [1, 4, 4, 9, 1],
-        ]
-    )
-    magazine.update(arr, robot_array)
-    arr = (
-        [
-            [2, 2, 2, 2, 3],
-            [1, "a", 1, "H", 3],
-            [1, "c", 1, "d", 3],
-            [1, "e", 1, "f", 3],
-            [1, 4, 4, 3, 3],
-            [1, 1, 9, 1, 1],
-        ]
-    )
-    magazine.update(arr, robot_array)
-    magazine.end()
-
-
-@pytest.mark.utils
-def test_reading_warehouse_map(warehouse_map1):
-    """
-    Test if ``utils.read_warehouse_map`` works properly.
-    """
-    map_ = [
-        [2,  2,  2,  2,  3],
-        [1, "a", 1, "b", 3],
-        [1, "c", 1, "d", 3],
-        [1, "e", 1, "f", 3],
-        [1,  4,  4,  4,  9],
-    ]
-    assert map_ == warehouse_map1
+from mmvdApp.shortest_path import neighbors
+from mmvdApp.shortest_path import a_star as shortest_path
 
 
 @pytest.mark.utils
@@ -114,6 +16,7 @@ def test_neighbors(warehouse_map1, position, expected):
                                  positions=True)
 
 
+@pytest.mark.utils
 def test_neighbors2(warehouse_map1):
     position = (0, 0)
     assert (0, 2, 1, 0) == neighbors(warehouse_map1, position,
